foundational-ai-server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foundational-ai-server"
version = "0.1.0"
authors = [
    { name="Aniket-think41" }
]
description = "Core package for Open Source CAI Pipecat - Foundational AI Server"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["ai", "pipecat", "open-source", "foundational-ai-server"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.115.6",
    "litellm==1.48.0",
    "livekit_api==0.8.0",
    "python-dotenv==1.0.1",
    "uvicorn==0.30.6",
    "alembic==1.13.3",
    "sqladmin==0.19.0",
    "SQLAlchemy==2.0.35",
    "sqlmodel==0.0.22",
    "asyncpg==0.30.0",
    "ruff==0.8.6",
    "deepgram-sdk==3.8.0",
    "pydantic_settings==2.7.1",
    "mcp==1.6.0",
    "pipecat-ai[cartesia,daily,openai,silero,webrtc,deepgram,websocket]==0.0.65",
    "pipecat-ai-small-webrtc-prebuilt==0.5.0",
    "av>=14.0.0",
    "websockets>=13.0.1",
    "openai-agents==0.0.11",
    "pydantic-ai[logfire]==0.0.51"
]

[project.urls]
"Homepage" = "https://github.com/think41/foundational-ai-server" # Please update if this is not the correct repository URL

[project.scripts]
"foundational-ai-server-cli" = "main:main"

[tool.setuptools]
py-modules = ["main", "models"]
packages = ["agent", "agent_configure", "custom_plugins", "utils"]
# MANIFEST.in will be used due to include_package_data=True being default or set by setup.py

